Use Search-F# to generate Fkey keycodes.
This is in addition to Search-#. On a Chromebook, it would only let you
generate F1 to F10. But on an external keyboard, it could let you generate
F1 to F12, and is more intuitive.

This also matches typical Fn-key behaviour on other keyboards with such a key.

R=yusukes@chromium.org,mazda@chromium.org,derat@chromium.org
BUG=164216

Review URL: https://chromiumcodereview.appspot.com/11437039

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@171560 4ff67af0-8c30-449e-8e8b-ad334ec8d88c