Add singleton argument to app.window.create().
If a window is given an id, trying to create a window with an id that is already
used will cause the existing window to be activated instead of a new window to
be created. You can override this behavior by passing singleton: false.

BUG=159217


Review URL: https://chromiumcodereview.appspot.com/11419295

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@171617 4ff67af0-8c30-449e-8e8b-ad334ec8d88c