Reorganize how messages are handled in metro ash (win8)
Before we had the mouse/keyboard messages cracked by the process host
which would blindly forward to the root windows host, now the
process host handles only one message and the input messages are
handled directly by the root window host.

Also the root window host now knows when it is connected
vs disconnected to the viewer.

This part 1 of a change which has the root window host send
messages to the viewer process.

BUG=151718
TEST= chrome ash works as before.
Review URL: https://codereview.chromium.org/11473005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@171648 4ff67af0-8c30-449e-8e8b-ad334ec8d88c