Rename Is*Ephemeral*() to Is*NonCryptohomeDataEphemeral()
When a regular user is logged in as ephemeral, both the content of the
user's cryptohome and any data stored outside the cryptohome (avatar,
wallpaper, etc.) are ephemeral.

For public account users, the content of the cryptohome is ephemeral but
data stored outside the cryptohome (avatar, wallpaper, etc.) should be
persistent.

This CL renames the UserManager's Is*Ephemeral*() methods to
Is*NonCryptohomeDataEphemeral() to make it clear that these indicate
what to do with data outside the cryptohome only.

There currently is no method to query whether the user's cryptohome is
ephemeral or not as that information is not needed anywhere.

BUG=158509


Review URL: https://chromiumcodereview.appspot.com/11443010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@171712 4ff67af0-8c30-449e-8e8b-ad334ec8d88c