Set up client-side support for having Finch trials enable the Instant Extended API.
Looks for a Finch trial of the name InstantExtended, with groups named Group[0-9]+.  Sets the &espv param equal to the group number.  If such a Finch trial exists, it is used in place of the value of the about:flags flag and the value of that flag is ignored.

Tested successfully with a local Finch server.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11439024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@171727 4ff67af0-8c30-449e-8e8b-ad334ec8d88c