Make MaybeWipeout function somewhat more efficient.
We only need to check the exclude prefs for bad extensions, not all extensions. Also, we don't need to check if the sideload wipeout has already happened each time we evaluate extensions (we can do that once and for all before we start evaluating extensions).

In other words, make an early return if the extension is of the wrong type/location before checking the prefs.

BUG=163697, 154624
TEST=None
Review URL: https://codereview.chromium.org/11439019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@171796 4ff67af0-8c30-449e-8e8b-ad334ec8d88c