[cros] Deferred animated resize of multiline user pod row.
Pods are moved and resized only when mouse cursor leaves the entire pod row.
This makes removing a bunch of users easy as the cursor stays on the remove button (of the next pod) after removing a user.
Podrow and pods resize is animated.


BUG=122346


Review URL: https://chromiumcodereview.appspot.com/11478023

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@171837 4ff67af0-8c30-449e-8e8b-ad334ec8d88c