Move ExtensionSystemImpl::extension_warning_service_ to ExtensionSystemImpl::shared
Move ExtensionSystemImpl::extension_warning_service_ to ExtensionSystemImpl::shared to prevent a crasher in Guest Mode of ChromeOS. The Guest Mode was a OTR profile that did not have a initialized ExtensionWarningService but was nevertheless used to generate the chrome://extensions page. By moving the EWS into a data container that is shared between the regular profile and OTR profiles (such as the Guest Profile) we fix this problem.


BUG=163007


Review URL: https://chromiumcodereview.appspot.com/11472020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@171957 4ff67af0-8c30-449e-8e8b-ad334ec8d88c