Properly disable accelerated compositing for extension background pages. Since these pages are not expected to ever render, there's no reason for the additional overhead.
In extension_webkit_preferences::SetPreferences() we were already trying to disable force_compositing_mode
but that wasn't taking effect as it was being reset later on in ChromeContentBrowserClient::OverrideWebkitPrefs.
I'm not sure why we would want to allow for any kind of GPU acceleration on background tabs so this patch
disables it all.

BUG=163698


Review URL: https://chromiumcodereview.appspot.com/11299288

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@172441 4ff67af0-8c30-449e-8e8b-ad334ec8d88c