Improve reliability of custom spelling dictionary file.
This improves reliability of custom spelling dictionary file through checksums,
backups, and atomic file writes.

When a user alters their custom dictionary, we back it up, calculate the
checksum of the new contents, and then write out the dictionary file with its
checksum atomically.

When the user restarts the browser and starts using spellchecker again, we load
the custom dictionary and check its checksum. If the dictionary checksum does
not match, but the backup checksum matches for the backup dictionary, we restore
the backup and use it instead.

BUG=161905

Review URL: https://chromiumcodereview.appspot.com/11414282

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@172494 4ff67af0-8c30-449e-8e8b-ad334ec8d88c