drive: Represent upload range in half-open interval [start, end), not [start, end].
In DriveUploader::UploadFileInfo, ResumeUploadParams, and ResumeUploadResponse,
this patch changes the range representation to be exclusive for |end_range|.
This makes less -1's and +1's in the high level code.

Conversion from/to the inclusive range required in the HTTP protocol
is carried out in the lowest level: ResumeUploadOperation.

BUG=163486

Review URL: https://codereview.chromium.org/11474052

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@172533 4ff67af0-8c30-449e-8e8b-ad334ec8d88c