This CL is part of https://codereview.chromium.org/11316275/, but only contains the changes related to the infobar.
The new design for the media infobar included in this patch:
# Use the confirm infobar for media, this will make the media infobar be consistent with other infobars in chrome like Geolocation.
# Remove the devices array in the media_stream_request since the infobar does not have a device dropdown any more.
# "Allow" will be sticky by default.
# Start using the device selection UI in the content settings.

Note, "Deny" is not sticky in this patch, it can't be done since the media code in content settings needs a big refactoring to handle the exceptions properly. It will be addressed in another CL.


BUG=163895
TEST=content_unittests, pyauto webrtc functional tests.
and manual tests like:
select a device via content setting media
go to https://webrtc-demos.appspot.com/html/pc1.html, and make a call.

Review URL: https://chromiumcodereview.appspot.com/11434084

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@172610 4ff67af0-8c30-449e-8e8b-ad334ec8d88c