SupportsUserData and manifest handlers for Extension; use them for the Omnibox API.
Note that manifest handlers aren't available in the renderer (there seems to be no place to register them). At least for Omnibox, the renderer doesn't need to parse the manifest field, but this won't be true everywhere.
Pull a bunch of Omnibox API stuff out of ExtensionService.

BUG=159265


Review URL: https://chromiumcodereview.appspot.com/11446034

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@172638 4ff67af0-8c30-449e-8e8b-ad334ec8d88c