google_apis:: DriveServiceInterface::GetAccountMetadata() returns AccountMetadataFeed
This way, client code doesn't have to convert base::Value to AccountMetadata

Note that the expectation in StaleCacheFilesRemoverTest.RemoveStaleCacheFiles
is updated. Previously, the mock version of GetAccountMetadata()  returned an
empty base::Value from the 2nd call, which was a bug. With this change,
GetAccountMetadata always returns a valid value to the caller.

BUG=165387
TEST=go to chrome:drive-internals and account metadata is shown

Review URL: https://codereview.chromium.org/11543014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@172796 4ff67af0-8c30-449e-8e8b-ad334ec8d88c