Remove partial file handling from DriveUploader.
Somewhile the class had dealt with the cases where the file to be uploaded
is not fully filled with the final contents (i.e. still being downloaded from
somewhere else). Since we have simplified the usecase and split the download
phase and upload phase, now we can safely assume the whole content of the
file is present at the time of UploadXXXfile().

Along the way, the patch removes some legacies (ununsed DEPS exception,
and IO thread in the test).

BUG=165110

Review URL: https://chromiumcodereview.appspot.com/11416362

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@172820 4ff67af0-8c30-449e-8e8b-ad334ec8d88c