New icons for Settings App in app launcher (and Windows taskbar).
This introduces a new designer icon for the Settings App with a
transparent margin so that it doesn't appear oversized. New assets
are also introduced -- high-quality pre-resized at 48x48px for the
icon in the app launcher, and 32x32 for the Windows 7 "large icons"
taskbar.

BUG=130210
TEST=Settings app icon updated to be slightly smaller, and now with
     8 cogs instead of 6.


Review URL: https://chromiumcodereview.appspot.com/11488025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@172846 4ff67af0-8c30-449e-8e8b-ad334ec8d88c