UserPolicySigninService now checks if a user is from a hosted domain.
Added UserInfoFetcher which makes an OAuth call to the userinfo service to fetch
information about the currently signed-in user.

UserPolicySigninServices uses a UserInfoFetcher to check whether a user is
from a hosted domain before trying to fetch a DMToken, to reduce the load on
the DMServer.

BUG=141123
TBR=brettw


Review URL: https://chromiumcodereview.appspot.com/11552019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@172894 4ff67af0-8c30-449e-8e8b-ad334ec8d88c