Ensure that desktop chrome AURA when launched on Windows 8, does not defer to an existing Chrome ASH instance.
The startup code which runs when a newly launched chrome exe defers to an existing chrome browser ends up activating the
new browser window in the last active desktop which was ASH if ASH is running. This is incorrect because the startup
code in chrome would never execute for an ASH instance.

We now default to HOST_DESKTOP_TYPE_NATIVE if there is no existing Browser instance passed to the
StartupBrowserCreatorImpl::OpenTabsInBrowser function

BUG=151718
R=robertshield
TBR=sky
Review URL: https://codereview.chromium.org/11576039

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@173236 4ff67af0-8c30-449e-8e8b-ad334ec8d88c