drive: Get rid of MockDriveWebAppsRegistry
The existing tests were just checking if MockDriveWebAppsRegistry::UpdateFromFeed()
was called, which was not a good test. We should instead check if the webapps
registry was properly loaded.

Besides, we can just use a real DriveWebAppsRegistry where MockDriveWebAppsRegistry
was used as a dummy (in stale_cache_files_remover_unittest.cc), as
DriveWebAppsRegistry is cheap to create.

BUG=166103
TEST=none

Review URL: https://codereview.chromium.org/11576031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@173284 4ff67af0-8c30-449e-8e8b-ad334ec8d88c