Add policy for limiting the session length
If this policy is set, the user is logged out when the time since login
reaches the limit. The CL adds a SessionLengthLimiter that enforces the
limit. A separate CL will add the UI that shows a countdown timer to the
user.

BUG=chromium-os:26957
TEST=Manual

TBR=sky@chromium.org
(vor chrome_browser_chromeos.gypi)

Review URL: https://codereview.chromium.org/11499012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@173285 4ff67af0-8c30-449e-8e8b-ad334ec8d88c