Implemented GetWindowSnapshot on RenderViewImpl
This necessitated the relocation of the previous chrome::GrabWindowSnapshot code
to ui/snapshot, which has been turned into it's own component to avoid circular dependencies with aura.
A new variant of GrabWindowSnapshot, GrabViewSnapshot, has been added as well to facilitate easier usage by views. chrome::GrabWindowSnapshotForUser was left in place to accomodate existing calls
to the API, but now calls up to the ui/snapshot code.

This is a subset of the prior CL 11362023, which has been broken apart to facilitate easier reviews

BUG=157479

Review URL: https://chromiumcodereview.appspot.com/11399002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@173329 4ff67af0-8c30-449e-8e8b-ad334ec8d88c