Fix search terms handling on commit.
The server calls setValue() with the query selected when a user clicks on a query suggestion.  When the page is in mode search, just update the omnibox text with the supplied query.

Also, fix what query we use in the VirtualURL on commit: use the user text if committing because of ENTER, and use the user text + the gray text if committing because of click on the preview.

Together, these should fix most of the flickering and wrong-text-in-the-omnibox issues. A better and more robust fix was started by Mathieu in crrev.com/11415292 but that approach will need some restructuring before it can land.

BUG=165710,159326

Review URL: https://chromiumcodereview.appspot.com/11570020

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@173526 4ff67af0-8c30-449e-8e8b-ad334ec8d88c