Add a base::ParseJson() function to help tests construct Values.
This is very similar to base::JSONReader::Read() except that it allows
trailing commas and EXPECT-fails with a good message when the json is
malformed.

I migrated two tests to use this as a demonstration that it's widely
useful and makes things more readable.


Review URL: https://chromiumcodereview.appspot.com/11567027

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@173540 4ff67af0-8c30-449e-8e8b-ad334ec8d88c