Resize ssl certificate selector dialog when hiding
In r173273 I introduced a bug where a the hidden SSL certificate selector dialog would still intercept mouse events.

To fix this I'm porting the solution used by GTMWindowSheetController. When hiding the sheet the window is resized to 1x1. When showing the sheet it's frame is reset to the old value.

BUG=166372


Review URL: https://chromiumcodereview.appspot.com/11614007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@173931 4ff67af0-8c30-449e-8e8b-ad334ec8d88c