Remove DCHECK for PSS::SetEncryptionPassphrase
For Android, we do not sync passwords, so the cryptographer is not
initialized on a non-encrypted account.

The |passphrase_required_reason_| is therefore set to REASON_ENCRYPTION
at the time of the DCHECK if you turn on custom passphrase encryption
after already being signed in and syncing data.

BUG=165309


Review URL: https://chromiumcodereview.appspot.com/11637045

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@174351 4ff67af0-8c30-449e-8e8b-ad334ec8d88c