Send information about all nodes which are touched by the user on the page from the renderer.
The ChromeViewHostMsg_FocusedEditableNodeTouched IPC has been renamed to ChromeViewHostMsg_FocusedNodeTouched with
a bool parameter indicating if the node is editable.

The NOTIFICATION_FOCUSED_EDITABLE_NODE_TOUCHED notification on the browser side has been renamed to NOTIFICATION_FOCUSED_NODE_TOUCHED
with a bool indicating if the node is editable.

This is done to be able to dismiss the On screen keyboard on Windows 8 if the user touches outside the editable field on the page.

BUG=166516
R=jam
Review URL: https://codereview.chromium.org/11636055

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@174426 4ff67af0-8c30-449e-8e8b-ad334ec8d88c