Split PrefService into PrefService, PrefServiceSimple and PrefServiceSyncable.
The first two (PrefServiceSimple is a subclass of PrefService) know
nothing about sync or any Chrome or content concepts.

The third (PrefServiceSyncable, a separate subclass of PrefService)
knows about sync and requires users to choose whether each individual
preference is syncable or not when it is registered.

BrowserProcess::local_state() is a PrefServiceSimple after this
change, and Profile::prefs() is a PrefServiceSyncable.

COLLABORATOR=kaiwang@chromium.org
TBR=ben@chromium.org
BUG=155525


Review URL: https://chromiumcodereview.appspot.com/11570009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@174531 4ff67af0-8c30-449e-8e8b-ad334ec8d88c