Close bubble when bookmark star button is clicked a second time
Currently when the bookmark star button is clicked a second time it closes and shows again.

The bookmark bubble already has logic to check if the bubble is showing and do nothing if asked to show again.

Unfortunately this doesn't work for mouse events because the bubble is closed on mouse down and shown on mouse up.

To fix this I'm adding a flag to check if the bubble was visible on mouse down. If the flag is set then the mouse up event is ignored.

BUG=107876
TEST=Clicked on the bookmark button a second time. Verified that the bubble didn't reopn.
Added a new browser test. Verified that the test failed without my change and passed with it.


Review URL: https://chromiumcodereview.appspot.com/11679008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@174846 4ff67af0-8c30-449e-8e8b-ad334ec8d88c