Fix regression: Inform UserManager::UserLoggedIn() when restarting after a crash
This CL fixes a mistake introduced in CL 11598015. When Chrome is restarting
after a crash, the browser_restart argument passed to UserLoggedIn() should be
|true|. Since browser tests (erroneously) use this code path as well, this
breaks a browser test that works only when the browser does not think it was just
restarted after a crash. The CL disables the test until the code path taken by
browser tests is corrected.

BUG=168057
TBR=ivankr@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11753002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@174921 4ff67af0-8c30-449e-8e8b-ad334ec8d88c