Public Accounts: Log out when the screen turns off on idle
This CL makes Chrome terminate a Public Account session when the user is
idle and the power manager turns off the screen. Without this CL, the
Public Account session would also terminate on idle but only after a
period of several minutes during which the screen is blank.

BUG=chromium-os:37593
TEST=Manual


Review URL: https://chromiumcodereview.appspot.com/11743018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@174989 4ff67af0-8c30-449e-8e8b-ad334ec8d88c