Windows: Remove desktop profile shortcuts (and any others pointing to the exe) on uninstall.
This is a re-land of https://codereview.chromium.org/11693010/, which failed on the tree on XP bots. The issue was that ProgramCompare::Evaluate() was parsing its value parameter is a command-line, which meant it was truncating it at the first space. This updated CL adds an EvaluatePath() function that takes a FilePath and doesn't try to parse it as a command-line. 

BUG=146636
TEST=Uninstalling Chrome removes any present profile shortcuts on the desktop.

Review URL: https://chromiumcodereview.appspot.com/11743022

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@175104 4ff67af0-8c30-449e-8e8b-ad334ec8d88c