Desktop aura: Expand what the ViewsDelegate can do to new windows.
(This is a reland of r173539, adding early bails to some implementations of OnBeforeWidgetInit so that we don't clobber an already set Widget::InitParams::native_widget.)

There are actual cases where we have to show dialogs with no parent or context,
when there are no browser windows. All of the cases that I can find go through
the CreateWindow() interface.

CreateWindow() now explicitly sets InitParams::top_level to true. The
ViewsDelegate now can do arbitrary manipulations on the InitParams so that when
top_level is true and we have no context, we either create a native_widget or
set the context based on desktop type.

BUG=165759
TBR=ben@chromium.org (who did the original review)

First Review URL: https://chromiumcodereview.appspot.com/11578014

Review URL: https://chromiumcodereview.appspot.com/11758002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@175131 4ff67af0-8c30-449e-8e8b-ad334ec8d88c