Implemented BluetoothExtensionFunction.
From discussion with Bryan (@bryeung) and Scott (@keybuk), we decided to make
the bluetooth extension functions asynchronous because it takes time to
initialize BluetoothAdapter. So I implemented BluetoothExtensionFunction, which
is the base class for all the subclasses and it calls DoWork() when the
bluetooth adapter is ready to use.

After this CL is submitted, I will turn
BluetoothAdapterFactory::DefaultAdapter() to receive a callback and run the
callback when the bluetooth adapter is ready.

TBR=sky@chromium.org
BUG=135470


Review URL: https://chromiumcodereview.appspot.com/11743024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@175553 4ff67af0-8c30-449e-8e8b-ad334ec8d88c