Don't pass DownloadItemModel ownership.
DownloadItemModel is a thin wrapper around DownloadItem*. It's
easier to instantiate it from a DownloadItem* than try to pass
ownership around.

Doing so was required when BaseDownloadItemModel existed, and the
actual model could be either a DownloadItemModel or a
SavePageModel. However this is no longer the case.

In addition, DownloadShelfContextMenu should track the lifetime
of the corresponding DownloadItem. It is possible for the
DownloadItem to be destroyed while the context menu is showing.

BUG=116551

Review URL: https://chromiumcodereview.appspot.com/11673004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@175558 4ff67af0-8c30-449e-8e8b-ad334ec8d88c