Revert 175099 > Omnibox: Make HistoryQuick provider not re-issue queries if |minimal_changes| is true > > BUG= > > > Review URL: https://chromiumcodereview.appspot.com/11722020
I'm reverting this because it causes problems with backspace, at least when
certain field trials are enabled.  See
https://code.google.com/p/chromium/issues/detail?id=168725

We'll probably chose a different strategy to fix this in a better way (in
terms of how the code should work), but at least this puts the behavior
back to what people are used to.

BUG=168725
TBR=mpearson@chromium.org
Review URL: https://codereview.chromium.org/11820015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@175668 4ff67af0-8c30-449e-8e8b-ad334ec8d88c