Add test for accessibility histograms.
The goal is just to make sure this code has test coverage, not to test
correctness. There was an issue in the past where the code that computed
the histograms crashed in debug mode, this will help prevent something like
that from landing in the future.

BUG=99504


Review URL: https://chromiumcodereview.appspot.com/11778045

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@175677 4ff67af0-8c30-449e-8e8b-ad334ec8d88c