Notify policy change observers via a task
This prevents unexpected reentrancy in OnUpdatePolicy() in the case that an observer changes policy (leading to a recursive call to OnUpdatePolicy()).

This avoids a crash that happens if we change the SyncDisabled policy, since
this signs out the user and modifies policy while we are still iterating through
the PolicyBundles in MergeAndTriggerUpdates().

BUG=155486


Review URL: https://chromiumcodereview.appspot.com/11667006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@175797 4ff67af0-8c30-449e-8e8b-ad334ec8d88c