Adds support for manual whitelist and blacklist, previewing a page and adding it to the whitelist for managed users.
This CL adds two user preferences (kManagedModeWhitelist and
kManagedModeBlacklist) which contain URL patterns which should be allowed
or blocked. The user preferences are used only to persistently store the
lists and the checking is done by the ManagedModeURLFilter, which uses
policy::URLBlacklist objects. The code takes care that the preferences and
the URLFilter are kept in sync all the time. It also contains the functions
needed to interact with these lists (adding, removing or checking).

It also adds a preview infobar which pops up when the user clicks Preview
on the interstitial. If the user clicks allow then the website is added
to the whitelist. The domain gets added plus all redirects that lead to
that page. The user is able to click navigate around the website without
getting another interstitial as long as he does not navigate to a different
host name.

BUG=168772

Review URL: https://chromiumcodereview.appspot.com/11299035

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@175806 4ff67af0-8c30-449e-8e8b-ad334ec8d88c