Distinguish plugin disconnections from plugin crashes.
If the channel between a renderer and a plugin encounters an error, currently we display an infobar saying "<plugin_name> has crashed.".
This may cause confusion, because the plugin process may still be alive and no crash dump is generated in that case.

With this CL, we display a different prompt if the plugin process is still alive:
"<plugin_name> has encountered an error."

This CL only affects Windows.

TEST=None
BUG=None


Review URL: https://chromiumcodereview.appspot.com/11820009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@176060 4ff67af0-8c30-449e-8e8b-ad334ec8d88c