drive: Add Profile* as a member of DriveProtocolHandler.
A Profile* can be passed to the constructor of DriveProtocolHandler and
the handler uses the given Profile* for processing the request.
This will help us adding unittests and introducting support for multiple profiles.
If the handler is constructed without Profile*,  it will use the profile from
ProfileManager::GetDefaultProfile().

BUG=167249
TEST=unittests. try opening a URL of 'drive://' (e.g. HTML file stored in Google Drive) on Chrome OS.


Review URL: https://chromiumcodereview.appspot.com/11785018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@176062 4ff67af0-8c30-449e-8e8b-ad334ec8d88c