Reset hidden property of butter-bar's progress-bar div when butter bar is invisible.
This is to stop computing animation in the butter bar's progress bar while the butter bar is not visible.
Just removing visible class from butter bar is not enough, because all that this
does is set the butter bar's opacity to 0, so the animation is still there,
it just becomes transparent.

BUG=165071
TEST=open Files app; copy a file and keep Files app window open;
verify the device eventually suspends after it's left idle


Review URL: https://chromiumcodereview.appspot.com/11801032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@176185 4ff67af0-8c30-449e-8e8b-ad334ec8d88c