[Autofill] Add Mac implementation for the in-browser process popup view.
There are still some TODOs, but the basic mouse-based functionality is implemented.

As additional context: The stable version of the Autofill popup UI is currently implemented in the renderer process -- actually, as part of WebKit.  It's implemented (cross-platform) as a heavily styled version of the <select> popup that we use on non-Mac platforms.  This CL implements this same popup as a native UI, to reach parity with GTK and Views.

BUG=51644

Review URL: https://chromiumcodereview.appspot.com/11740033

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@176558 4ff67af0-8c30-449e-8e8b-ad334ec8d88c