[Mac] Remove autofill popup window before closing.
For some reason, on my MBP with retina display, closing the child
window while it still a child causes the parent window to be ordered
out.  Switching apps and switching back brings it back, but the child
window is now pinned and cannot be removed without quitting.

BUG=169933

Review URL: https://codereview.chromium.org/11885032

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@176774 4ff67af0-8c30-449e-8e8b-ad334ec8d88c