Rearchitecting the 'app mode' code on mac for v2 apps.
A lot of the old code assumed an outdated model for v2 packaged apps on mac;
specifically that they used to run in their own Chrome process with their own
user data directory. This is no longer the case, and as such it's not possible
to have apps running in their own process, because they need access to profile
information from Chrome.

This patch changes the code so that the 'app mode' .app bundles function more
like shortcuts to open the app in an existing Chrome browser process (launching
Chrome if it's not already running). This allows the app to be launched via
Spotlight and the /Applications folder, and the user can e.g. drag the app onto
the dock like a regular Mac application.

There is as yet no way to trigger the shortcut creation code.

BUG=168080


Review URL: https://chromiumcodereview.appspot.com/11737014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@176803 4ff67af0-8c30-449e-8e8b-ad334ec8d88c