Renderer form_autofill_util changes to support Checkboxes and Radiobuttons.
Checks for checkable elements while reading the loaded page.

This is attempt no. 3 to submit this change in. Attempt no. 2 failed as it regressed on win7-* builds, while it was good on linux-release builds. This CL is one of the smaller CLs created out of r176189 in an attempt to narrow down the code effecting win7-* builds.


BUG=157636


Review URL: https://chromiumcodereview.appspot.com/11875019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@176829 4ff67af0-8c30-449e-8e8b-ad334ec8d88c