Add base class cr.ui.BubbleButton for options.ControlledSettingIndicator
Re-landing. The test failure this was reverted for is unrelated (the
failure was in native UI, this is a WebUI change).

This CL extracts a cr.ui.BubbleButton base class from the
options.ControlledSettingIndicator so that it can be shared with the
upcoming improved about:policy page.

BUG=134849
TEST=Manual

TBR=jhawkins@chromium.org


Review URL: https://chromiumcodereview.appspot.com/11907007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@176946 4ff67af0-8c30-449e-8e8b-ad334ec8d88c