Fix for a crasher when starting a renderer.
A crash dump from the field reports that the CrashDumpManager in
ChromeContentBrowserClient is not initialized when
GetAdditionalMappedFilesForChildProcess is called.
This is because the CrashDumpManager is initialized when a RenderProcessHost
is created, but it could be the case that a BrowserChildProcessHost gets
created first and calls GetAdditionalMappedFilesForChildProcess before
any RenderProcessHost was created.
Making the CrashDumpManager a singleton initialized early on to prevent this
crasher.

BUG=163456
TEST=None


Review URL: https://chromiumcodereview.appspot.com/11884042

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@177110 4ff67af0-8c30-449e-8e8b-ad334ec8d88c