Pepper: Introduce ThreadAwareCallback.
Some PPB interfaces have methods that set a custom callback. Usually, the callback has to be called on the same thread as the one it was set on. ThreadAwareCallback keeps track of the target thread, and posts a task to run on it if requested from a different thread.

BUG=None
TEST=newly added unittests.


Review URL: https://chromiumcodereview.appspot.com/11859015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@177319 4ff67af0-8c30-449e-8e8b-ad334ec8d88c