Use Manager.UninitializedTechnologies to track initializing state for Cellular
It turns out that we were unable to correctly use EnabledTechnologies + Device.Powered to determine the Cellular Initializing state.
Instead, we are adding an "InitializedTechnologies" property. For now, this is a shared string, but we'll move it to service_constants.h before we commit to ToT.

BUG=169979


Review URL: https://chromiumcodereview.appspot.com/11971009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@177342 4ff67af0-8c30-449e-8e8b-ad334ec8d88c