This CL adds more api's to developer_private to enable / disable extensions, load unpacked, uninstall, allow file access etc.
It also adds the related UI code to display the apps.

There are few API's which are still to be implemented like
Pack-extension, allow-incognito etc.

I will add them in the following CL. Also, browser tests
to follow in the follow up CL.



More APIs for AppsDebuggerPrivate.

The AppsDebugger UI code is in 12021003


BUG=149036
TBR=sky@chromium.org
TBR=estade@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11734031

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@177648 4ff67af0-8c30-449e-8e8b-ad334ec8d88c