Show app launcher after install an app, if the app launcher is installed.
THe app launcher was previously shown when the launcher was enabled via a flag, but not when it was enabled by being installed.

This change introduces a platform independent way of determining if the app launcher is enabled. This will also be used in the New Tab Page.

TBR=sky@chromium.org for gypi change
BUG=152854


Review URL: https://chromiumcodereview.appspot.com/11975035

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@177656 4ff67af0-8c30-449e-8e8b-ad334ec8d88c