Add GetRootResourceId method to DriveServiceInterface.
The resource id for WAPI is "folder:root" but one for Drive API is "root".
To fill the gap, this CL adds the new method GetRootResourceId, so that clinets
don't need to use hard-coded parameters.
This is preparation to replace content URL with resource id for directory
operations.

BUG=148627, 148628, 148629
TEST=Ran unit_tests.


Review URL: https://chromiumcodereview.appspot.com/11968038

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@177893 4ff67af0-8c30-449e-8e8b-ad334ec8d88c