drive: Deal with the root directory same as others in MoveOperation.
Our current implementation of file moving is done in the following logic:
"if(source!=root) RemoveFrom(source); if(target!=root) AddTo(target)"
implicitly assuming "belonging to nowhere" means belonging to root.
But this is not the case.

This patch removes the "if !=root" checks and treats the root as same as other directories.

BUG=169420
TEST=Manually test the steps in the bug report.

Review URL: https://chromiumcodereview.appspot.com/12039005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@177997 4ff67af0-8c30-449e-8e8b-ad334ec8d88c