History: Add option to group visits by domain
This patch is part of a series of 3 patches that add grouping by domain name
to the Chrome history page.

The current patch adds checkbox to the history page, which, when enabled,
groups the results in a day by domain and orders them by the number of visits
to that domain.

The change is only active when the "--enable-grouped-history" command line
switch is passed when running Chrome.

BUG=170690
TBR=jhawkins@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11975053

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178007 4ff67af0-8c30-449e-8e8b-ad334ec8d88c