Do not update the geolocation permission for local pages
When chrome is launched with --disable-web-security and a local
page tries to use the geolocation api, the chrome browser will ask
permssion from user and then will update geolocation permission for
the local page with URL file:///. But the file:/// is not a valid
ContentSettingsPattern. It will cause crash in debug build. This change
fixes the crash.

BUG=171323

Review URL: https://chromiumcodereview.appspot.com/12033015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178026 4ff67af0-8c30-449e-8e8b-ad334ec8d88c