Enable breakpad building by default on Android.
After https://chromiumcodereview.appspot.com/10407058 we can compile
breakpad by default and still have it disabled for non-official builds.

Changes the Android build to allow compiling breakpad but not use
it by not creating the crash fd and not passing it to the renderer
process unless breakpad is enabled.

Changes linux and Android to use a switch for enabling breakpad
since that's a lot easier to test with on android then an environment
variable.

BUG=105778,170530


Review URL: https://chromiumcodereview.appspot.com/11969025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178111 4ff67af0-8c30-449e-8e8b-ad334ec8d88c