Increased notification close button click area & added state images.
Expanded the clickable area for notifications from 8 x 8 pixels to
29 x 29 pixels and added hover and pressed state images.

Implementing this required moving the close button from a grid layout
column to a separate layer of the notification view, which caused a
number of changes in NotificationView's drawing.

In passing, references to one of the notification types the notification
view supports were changed from "multiple-item notifications" to the new
(and improved :-) "list notifications".

BUG=168822


Review URL: https://chromiumcodereview.appspot.com/11946035

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178196 4ff67af0-8c30-449e-8e8b-ad334ec8d88c