Disable VisitedLink in incognito mode
Fix regression introduced in r175906 where the default and incognito
profile shares the same VisitedLinkMaster. Before r175906, a separate
VisitedLinkMaster is created for incognito profile but since no
HistoryService is created for the incognito profile, the
VisitedLinkMaster never has urls added to it.

In this CL, just skip creating an VisitedLinkMaster for the incognito
profile in the first place.

BUG=171374


Review URL: https://chromiumcodereview.appspot.com/12039025

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178245 4ff67af0-8c30-449e-8e8b-ad334ec8d88c