<webview>: Implement ExecuteScript
This patch implements executeScript for <webview> by using extensions bindings for forwarding requests to the browser process from the app process.

The <webview> shim passes the ProcessId and the RouteID of the guest process to the ExecuteScriptFunction object in the browser process. From there, ExecuteScriptFunction grabs the guest web contents, and creates a ScriptExecutor object attached to the guest WebContents.

The callback is supported trivially through the extension bindings.

When a new guest web contents is created, we inject the Chrome App's extension information into the guest process, so that executeScript knows to bypass permission requests when attempting to execute script within the guest (as suggested by mpcomplete@).

BUG=153530
Test=WebViewTest.Shim, webViewExecuteScript

Review URL: https://codereview.chromium.org/11968054

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178520 4ff67af0-8c30-449e-8e8b-ad334ec8d88c