Fixed navigation in mosaic view using up/down keys (Files.app).
In mosaic mode, when more than one raw is visible, and tiles are not the same size, then up/down arrow keys' behavior was faulty. This patch solves this issue by overriding getIndexAbove and getIndexBelow methods on the selection controller.
Before, up/down keys were moving to the previous or to the next item. Now, they move to an item optically above or below.

TEST=Enter Files.app, gallery, mosaic view, lots of pictures, big window to have multi rows. Use up/down keys.
BUG=170615


Review URL: https://chromiumcodereview.appspot.com/12041013

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178585 4ff67af0-8c30-449e-8e8b-ad334ec8d88c