Extensions that use WebRequest API should not interact with <webview>
This patch exposes the flag that indicates whether a render process is a guest
to the IO thread in the browser process. When the WebRequest API checks for
matching listeners, it checks if the request is coming from a <webview> guest
process and if so, it checks if the listeners have a target process/routing ID
(no listeners currently do). Thus, the current WebRequest API no longer
interacts with <webview>.

Upcoming patches will enable the WebRequest API for guests via a slightly
different API.

BUG=171421
Test=Manually

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=178528

Review URL: https://chromiumcodereview.appspot.com/12052014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178651 4ff67af0-8c30-449e-8e8b-ad334ec8d88c