Integrate out-of-process iframes prototype with the Task Manager.
I'm adding a method to the RenderViewHost to indicate whether it is used for rendering a subframe. In the long term, we will need such a method for distinguishing top level renderers versus out-of-process iframe renderers, so I'm introducing it now, even though it might not be the final form.
The task manager is also modified to use this method to properly display iframes from different processes.

TBR=yoshiki@chromium.org

BUG=99379

Review URL: https://chromiumcodereview.appspot.com/12052056

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178710 4ff67af0-8c30-449e-8e8b-ad334ec8d88c