Instant should have a dedicated render process
This change introduces the notion of an Instant-specific render process. The
intent being that Instant-related sites can be grouped so as to share common
priveleges and be isolated on a process-per-site basis.

It introduces a new "chrome-search:" scheme, for use in grouping Instant URLs.

It adds methods to the template_url.* code to determine whether a given url is
an Instant url.

This paves the way for Instant-specific enablements, such as the SearchBox V8
extension, and theme data access for Instant pages.

BUG=168202
TEST=TemplateURLTest.*, InstantTest.ProcessIsolation
R=sreeram@chromium.org, creis@chromium.org
TBR=pkasting@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11782021

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178755 4ff67af0-8c30-449e-8e8b-ad334ec8d88c