Resubmitting asynchronous bluetooth adapter init CL with fix.
The original CL: https://chromiumcodereview.appspot.com/12018024/ was reverted because it violated the style guide that the struct-level global variable is prohibited and it failed the static analyzer test, in bluetooth_adapter_factory.cc. In order to resolve the issue, I changed adapter_callbacks to a lazy instance.

BUG=135470

Review URL: https://chromiumcodereview.appspot.com/12045051

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@178798 4ff67af0-8c30-449e-8e8b-ad334ec8d88c