Temporarily remove TRACE argument on ValueStoreFrontend::Backend as it uses a stale pointer
TRACE_EVENT captures const char* by value and only serializes those const
char*'s when the trace ends. As coded, this causes the trace to contain memory
garbage which prevents loading. The proper fix is to turn the value into a std::string, which
the tracing system will copy.

I did not do that change because it wasn't immediately obvious how to coerce
this into a string.

TBR=jeremy@chromium.org

Review URL: https://codereview.chromium.org/12093010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@179092 4ff67af0-8c30-449e-8e8b-ad334ec8d88c