Add sign-in promo to the Extension Installed bubble (mac).
Overview of xib changes (see screenshot in bug):
Added a TextField placeholder for a sync sign-in promo.
Widened the TextFields just a bit as there looked like there was a bit too much whitespace to the right of them (compared to other sides of the bubble).

BUG=164955
TEST=Automated tests.
Review URL: https://codereview.chromium.org/11997002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@179139 4ff67af0-8c30-449e-8e8b-ad334ec8d88c