Add backup URL support for SafeBrowsing data requests.
If a request fails, an alternate URL is tried immediately. This is
only done if the configuration provides backup URLs to try. 

Three cases of request failure are supported:
  - Issues with DNS resolution, TCP connection, or SSL handshake when
  the client is believed to be connected to the internet.

  - Issues with invalid HTTP responses from the server.

  - Issues with general network connectivity from the client.

If the backup request fails as well, then the entire update is treated
as a failure and no further backups are attempted.
 
BUG=155753
TEST=unit_tests --gtest_filter=SafeBrowsingProtocolManagerTest.*


Review URL: https://chromiumcodereview.appspot.com/11784038

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@179412 4ff67af0-8c30-449e-8e8b-ad334ec8d88c