Separate local and remote sync invalidations
The old design used ChromeSyncNotificationsBridge to register for both
NOTIFICATION_SYNC_REFRESH_LOCAL and NOTIFICATION_SYNC_REFRESH_REMOTE.
The first of these is issued when some component decides that the syncer
ought to perform an unsolicited GetUpdates for some type, while the
second is an Android-specific mechanism for receiving remote
invalidations.  In either case, the ChromeSyncNotificationsBridge would
forward the invalidations to the SyncManager through the
BridgedInvalidator as though they had come from any other invalidator.

In this new design, the local and remote invalidations sources are kept
distinct.  The SyncBackendHost is now responsible for listening to and
forwarding NOTIFICATION_SYNC_REFRESH_LOCAL.  Note that there is no
longer any type-specific filtering performed as we receive
notifications; it will be up to the SyncManager to ignore notifications
for disbaled types.

The successor to ChromeSyncNotificationsBridge,
AndroidInvalidatorBridge, is responsible only for forwarding
NOTIFICATION_SYNC_REFRESH_REMOTE notificaitons.  It is passed to the
SyncManager only if OS_ANDROID is defined.

Since there is only one source of remote invalidations, the
BridgedInvalidator would seem to be no longer necessary.  Unfortunately,
this is not the case.  The BridgedInvalidator was useful in part because
it wrapped the UI-thread-owned ChromeSyncNotificationsBridge without
owning it, so the SyncManager's destruction would not affect it.  This
wrapping functionality is still needed by the AndroidInvalidatorBridge.
The BridgedInvalidator has been modified to only wrap a single
AndroidInvalidatorBridge, and renamed to AndroidInvalidatorBridgeProxy.

BUG=124143




Review URL: https://chromiumcodereview.appspot.com/12022041

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@179546 4ff67af0-8c30-449e-8e8b-ad334ec8d88c