Introduce BrowserIterator. An iterator meant to iterate over all the browsers in BrowserListImpls across multiple desktops (i.e. multiple HostDesktopTypes).
As a first step, used it in TabContentsIterator (pretty much bringing back the implementation to the one that was there prior to https://codereview.chromium.org/11316340/ -- except that the iterator is now aware of multiple desktops).

Another CL will follow in which all BrowserList::const_iterator, etc. methods will be replaced by BrowserIterator.

The interface of BrowserIterator is derived from the interface of TabContentsIterator for consistency (which is being slightly tweaked in parallel in https://codereview.chromium.org/12049038/).

BUG=129187

Review URL: https://chromiumcodereview.appspot.com/12038053

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@179695 4ff67af0-8c30-449e-8e8b-ad334ec8d88c