Alternate NTP: Add search token to omnibox
This is the Mac verison of this Views CL:
https://chromiumcodereview.appspot.com/11418229/

This CL adds a search token decoration to the omnibox.

The search token is displayed when the user does a search from the omnibox.

The search token appears on the right with a separtor between it and other decorations.

Screenshot:
  no separator: http://i.imgur.com/TXCixtn.png
  with separator 1x: http://imgur.com/xtxv9hG
  with separator 2x: http://imgur.com/2Pzi7dN

BUG=163190
TBR=thakis@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12042002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@179818 4ff67af0-8c30-449e-8e8b-ad334ec8d88c