iOS and non iOS use different decoder for SkBitmap to PNG. Currently: - Sync sends the favicon PNG data to the history database - The history database sends a notification (indirectly) to sync that its favicon data has changed.   In the process of the 'updated data' being sent to sync,   the PNG data is decoded to ImageSkia and reencoded to PNG.   Sync receives the reencoded data.
Sync compares byte-for-byte the 'updated data' with the data that it has sent to the database.
If the data from sync was encoded by a different PNG codec than the one used natively, the 'updated data' doesn't match exactly and sync erroneously thinks that it has an update to push to the server.

This CL removes the decode/encode if possible in in FaviconUtil::SelectFaviconFramesFromPNGs to avoid pushing erroneous updates.

BUG=172053, 171465
Test=SingleClientBookmarksSyncTest.SetFaviconHiDPIDifferentCodec

Review URL: https://chromiumcodereview.appspot.com/12034002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@179974 4ff67af0-8c30-449e-8e8b-ad334ec8d88c