Allows to show display settings for 1-display of stumpy.
The old code just checks the number of displays and shows the settings
only when there're 2>= displays. However we may want to provide the
display settings when only one display is connected to stumpy. That's
because the user may want to calibrate the display overscan for the
single display.

In the new code, the display settings is visible only when there is
at least one external display, so the behavior is same on chromebooks
but the settings will be visible on stumpy.

BUG=172936
TEST=Display settings is visible on chromeos-chrome on Goobuntu

Review URL: https://codereview.chromium.org/12088081

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@179982 4ff67af0-8c30-449e-8e8b-ad334ec8d88c