Remove syncer::IncomingInvalidationSource
Since we separated local and remote invalidations in r179546, we no
longer need to differentiate between the two sources with an enum.  The
IncomingInvalidationSource is always REMOTE_INVALIDATION.

This commit removes the IncomingInvalidationSource enum and the
'source' parameter that was part of the invalidator callback API.  The
callees can now safely assume that all invalidations have a 'remote'
source.

BUG=124143


Review URL: https://chromiumcodereview.appspot.com/12114012

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@180213 4ff67af0-8c30-449e-8e8b-ad334ec8d88c