Refactor: Simplify WaitableEventWatcher.
This change uses a callback instead of a delegate for specifying what
should be called when a WaitableEvent occurs.

This simplifies the class and gets rid of a workaround internal to the
class to prevent name collision on "Delegate" inner classes.

Tested (linux and windows):
  out/Debug/base_unittests --gtest_filter=*WaitableEventWatcherTest*
  out/Release/ipc_tests with valgrind and leak-check=yes

Previously reverted at:
  https://codereview.chromium.org/12087120/


BUG=


Review URL: https://chromiumcodereview.appspot.com/12094106

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@180450 4ff67af0-8c30-449e-8e8b-ad334ec8d88c