Chrome: Enable ChromeOS-Wide-Profiling
We want to gather perf data collected by running "perf record -a" from opt-in
ChromeOS users. This CL enables that.

We use debugd in ChromeOS to run quipper, a program that runs "perf record -a"
and returns that data in a protobuf. We do this every 12 hours for 2 seconds.
The overhead is minimal (5% slowdown for 2 seconds on "lumpy" machines).

BUG=157508
TEST=Manual:
Change the perf record interval from 12 hours to 1 minute and set a
breakpoint on ParseProtoIfValid(). It gets hit and the protobuf is parsed
properly.


Review URL: https://chromiumcodereview.appspot.com/11185038

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@180484 4ff67af0-8c30-449e-8e8b-ad334ec8d88c