Move DownloadFileOperation to base_operations.
The operation can be shared between GData WAPI and Drive API v2. So, as the
preparation of the migration, this moves it and its test to base_operations.
Note that the unittest for the moved class is in base_operation_server_unittest.cc. The reason why it is split from the existing base_operation_unittest.cc is that unfortunately the test won't work on Android. The cause is as same as gdata_wapi_operations_unittest/drive_api_operations_unittest. So, disabled it on Android in the same way with them.

BUG=148627
TEST=Ran unit_tests.

Review URL: https://chromiumcodereview.appspot.com/12088104

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@180628 4ff67af0-8c30-449e-8e8b-ad334ec8d88c