App Shortcuts on Linux now have icons at full size, not 32x32.
The create application shortcuts dialog no longer loads the icon at 32x32 for
display in the dialog and saving to the shortcut. Instead, it loads the icon at
full size, and scales it to 32x32 just for display. It saves at the largest
possible size (up to 1024, since icon loading requires an upper limit).

Also enforce a minimum width for the description label in that dialog. (This
doesn't affect current behaviour, but makes the dialog robust against larger
icon sizes in the display.)

BUG=173950


Review URL: https://chromiumcodereview.appspot.com/12179009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@180665 4ff67af0-8c30-449e-8e8b-ad334ec8d88c