History: Add range navigation control for grouped visits
The current patch enables the user to select the time frame
for which the the grouped results are viewed by
adding a set of buttons. These buttons allow the user to
go back in time, reset to today or go forward until the current interval.

BUG=170690

Review URL: https://chromiumcodereview.appspot.com/11886104

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@180715 4ff67af0-8c30-449e-8e8b-ad334ec8d88c