Add synchronous Widget closing event
Adds a WidgetObserver event that is guaranteed to be invoked synchronously
when a Widget is requested to close.  Rename the existing event, which
is invoked at Widget destruction time, typically via a task on the event
loop, to OnWidgetDestroying.

The new event will be used in Views-specific WebContentsModalDialogManager
functionality to get close events for Widgets, so that we don't risk
adverse interleaving of Widget close events with other relevant events
(e.g. IPC).  This will allow us to avoids test flakiness observed
in https://codereview.chromium.org/12045037 due to event races in
PrintPreviewTest.PrintCommands

BUG=157161


Review URL: https://chromiumcodereview.appspot.com/12089061

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@180796 4ff67af0-8c30-449e-8e8b-ad334ec8d88c