Migrate invalidator client IDs out of sync
We want to store invalidator client IDs somewhere other than the sync
DB.  Unfortunately, changing in the ID would require us to clear any
existing client state, which is something we'd prefer to avoid.

This commit makes the SyncInvalidationListener copy its sync-assigned
client ID to the InvalidationStateTracker when it is initialized.  When
we decide to fully separate the sync and notifier IDs, the clients that
have run through this code path will find that their notifier client ID
has already been initialized correctly.  This allows us to avoid
clearing their client state.

It's impossible to guarantee that all clients will hit this code path.
Fortunately, we don't have to.  The goal of this patch is to avoid
deleting the state of too many clients all at once.  It's not a big
problem if we have to clear the state of a few stragglers that haven't
hit the migration path yet.

BUG=124142


Review URL: https://chromiumcodereview.appspot.com/12094113

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@180907 4ff67af0-8c30-449e-8e8b-ad334ec8d88c