Implement user consent InfoBar for screen capturing.
Support for screen capturing using getUserMedia() was added 
in r180502 (behind a flag), but there were no notifications 
to let the user know that the screen is being captured. This 
CL dds InfoBar that is shown whenever a page requests screen 
capturing and lets the user approve or deny the request.

BUG=134249
TBR=brettw@chromium.org (gypi changes)

Review URL: https://chromiumcodereview.appspot.com/12022024

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@181112 4ff67af0-8c30-449e-8e8b-ad334ec8d88c