Clicking on the caption in desktop chrome AURA eats subsequent mouse events until we click away from the caption.
When we receive a mouse pressed event in AURA we set the mouse pressed handler and set it back to NULL when we receive a mouse
released event.

On Windows clicking on the caption sends in the WM_NCLBUTTONDOWN message. However we don't receive the subsequent WM_NCLBUTTONUP message.
As a result the mouse pressed handler remains non NULL and ends up eating all mouse messages until we click away from there. For e.g
on the omnibox or on the page.

This works on non AURA chrome by chance as we ignore the mouse down on the caption.

Fix as per discussion with sky is to not set the mouse event handler for non client mouse downs.

I also added the WM_NCMOUSELEAVE message to the list of mouse messages not sent to the renderer in the render_widget_host_view_aura.cc file.
Was seeing an annoying assert on hovering on the scrollbar.

BUG=174527
TEST=Covered by interactive ui test MouseDownOnBrowserCaption
Review URL: https://codereview.chromium.org/12239002

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@181142 4ff67af0-8c30-449e-8e8b-ad334ec8d88c