On Linux, Chrome App windows now set their window class to match the StartupWMClass field of the app shortcuts.
This means on Ubuntu Unity (and potentially other platforms that group windows
into applications), if an application has a shortcut installed, its windows will
use the correct (full resolution) app icon and app title. Also, on Unity, apps
can be locked to the launcher and re-launched from the launcher icon.

Technical details: The WM_CLASS of app windows used to be the title of the app;
it is now the string "crx_" followed by the app ID, which matches the
StartupWMClass string set in the .desktop file, and also avoids collisions if
two apps have the same title.

BUG=173314


Review URL: https://chromiumcodereview.appspot.com/12221037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@181194 4ff67af0-8c30-449e-8e8b-ad334ec8d88c