Extract UploadRangeOperationBase class from ResumeUploadOperation.
The newly being introduced class GetUploadStatusOperation looks very similar
to ResumeUploadOperation, especially the access url, http method, and how to
handle the response from the server. The diff should be the request header
and body (content) (along with it, progress update).
As a preparation of the implemetation of GetUploadStatusOperation,
the base class is extracted.

BUG=148632
TEST=Ran unit_tests


Review URL: https://chromiumcodereview.appspot.com/12224026

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@181279 4ff67af0-8c30-449e-8e8b-ad334ec8d88c