Split InitiateUploadOperation for new file upload and existing file upload.
This CL splits InitiateUploadOperation into two classes,
InitiateUploadNewFileOperation and InitiateUploadExistingFileOperation.
The main goal is geet rid of a hack of |upload_location|.
In the current code, |upload_location| is parent directory's URL if
uploading a new file. However, if uploading an existing file, it is the file's
URL. This is the preparation to replace the URL to resource id.
From the implementation point of view, this CL extracts the common code for
those two class, and creates a new base class named InitiateUploadOperationBase.
The class can also be used to implement initiate uploading on Drive API v2,
so it is in base_operations files instead of gdata_wapi_operations files.

BUG=148632
TEST=Ran unit_tests


Review URL: https://chromiumcodereview.appspot.com/12208069

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@181464 4ff67af0-8c30-449e-8e8b-ad334ec8d88c