Refresh Drive files' available-offline status in File Manager on time.
This CL does two things:
(1) Notify OnDirectoryChanged when a file is stored to the cache.
(2) Fix JS side logic to deal with the case when the available-offline bit is set.

For (1), the JS side does some frequency control (directory_model.js) so
it shouldn't cause bad flickering in UI etc. For (2),
it is simply a bug and hence should be fixed.

BUG=174458

Review URL: https://chromiumcodereview.appspot.com/12210079

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@181469 4ff67af0-8c30-449e-8e8b-ad334ec8d88c