Make devtools_window.cc multi-desktop aware.
Getting rid of BrowserList:: methods which only know about the native-desktop.

Note that CreateBrowserForDevTools is still native desktop only so that devtools will not fully work yet in a multi-desktop environment, but this is making progress in the refactoring goal of getting rid of all native desktop only methods.

BUG=129187


Review URL: https://chromiumcodereview.appspot.com/12210063

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@181697 4ff67af0-8c30-449e-8e8b-ad334ec8d88c