Alternate NTP: Fix dev tools layout
When docked to the right the dev tool's web view would be cropped.

This bug was due to the way I implemented view layout in r181301.

The dev tool view was being correctly layed out but the subview was not.

To fix this I've removed the container view (DevToolsContainerView). The new code just use the view's auto layout instead.

BUG=173205
TEST=Docked the devtools to the right and opened a new browser window. Verified that the devtools view was correctly layed out.

Review URL: https://codereview.chromium.org/12226086

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@181867 4ff67af0-8c30-449e-8e8b-ad334ec8d88c