Remove unused lock, add verification of intended usage.
There used to be locking in this class, but many small changes had contributed to removing all use of the lock. The lock does not appear to actually be needed, as long as the class continues to be used as it currently is used, with the deferred initialization via RegisterExtensionService always happening before the providers map owned by the object is iterated over.

This change removes the |lock_| member and adds debug-only code that verifies that this is how the class continues to be used.

BUG=none

Review URL: https://codereview.chromium.org/12223052

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@181930 4ff67af0-8c30-449e-8e8b-ad334ec8d88c