Fixed unexpected scroll when removing profiles in settings.
The problem was caused by auto focusing to the first focusable control when switching pages. To show an overlay, a owning page is shown, very often to the same current one, what is causing setting focus to the first focusable item, and scrolling to the top just before showing an overlay. This patch addresses this issue by not changing focus if the current focused control is on the active page/overlay.

TEST=Follow the bug's repro steps.
BUG=174608

Review URL: https://codereview.chromium.org/12226110

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182133 4ff67af0-8c30-449e-8e8b-ad334ec8d88c