Fix crash around Reset function call.
There is timing issue around initialization, if the InputContext is initialized
after InputMethodManager, Reset function will be called with NULL object proxy.
This cause a crash. If the object proxy is not ready, calling Reset is not
necessary. So just doing nothing is fine.

BUG=174248
TEST=None


Review URL: https://chromiumcodereview.appspot.com/12223084

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182147 4ff67af0-8c30-449e-8e8b-ad334ec8d88c