Layout the NTP footer with flexboxes.
This updates the CSS flexbox property names to their most recent
iterations, displays the entire footer in a flexbox to make the layout
more obvious, and rearranges the footer's right-side flexbox so that it
always has the same width.

The logo on the footer's left side is resized once when the page is
rendered to make it the same width as the right side.  This keeps the
nav dots centered in the footer.

BUG=


Review URL: https://chromiumcodereview.appspot.com/12038090

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182190 4ff67af0-8c30-449e-8e8b-ad334ec8d88c