Fix issue 174250, crash when reloading packaged app.
When a shell window was closed, it wasn't removed from the registry immediately
(it was removed in a callback from OnNativeClose, which happens a bit later.)
This meant that when the app was loaded again, the window was still in the
registry with a stale pointer to the old extension. ExtensionSettingsHandler
was tickling this by iterating shell windows on extension load.

BUG=174250


Review URL: https://chromiumcodereview.appspot.com/12210107

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182194 4ff67af0-8c30-449e-8e8b-ad334ec8d88c