Omnibox: Possibly Fix Rare OpenMatch Crash
The associated bug seems to imply a crash on this line
ClassifyPage(controller_->GetWebContents()->GetURL())
in OmniboxEditModel::OpenMatch().

This change corrects the part of the test that may fail
(if there's no active tab for some reason).
I think controller_ is always valid (the rest of the edit
model code use it without testing it for non-NULL).
But I see at least two places in OmniboxEditModel that
test for the non-NULLness of controller_->GetWebContents().
Hence, I think it might be able to be NULL, so this
change adds a test for it here too.

TEST=new code compiles
BUG=142931


Review URL: https://chromiumcodereview.appspot.com/12250014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182244 4ff67af0-8c30-449e-8e8b-ad334ec8d88c