Ensure that when themes in desktop chrome AURA on Windows use the opaque browser frame view.
To achieve this we need to mimic the implementation in BrowserFrameWin::ShouldUseNativeFrame in the
DesktopBrowserFrameAura::ShouldUseNativeFrame override.

The code in BrowserFrameWin::ShouldUseNativeFrame has been moved to a helper function in the browser namespace
ShouldUseNativeFrame in the newly added files browser_frame_common_win.h/.cc. This is invoked from BrowserFrameWin
and from the BrowserDesktopRootWindowHostWin::ShouldUseNativeFrame override which is invoked via the
DesktopBrowserFrameAura::ShouldUseNativeFrame override for AURA

BUG=175372
R=sky
Review URL: https://codereview.chromium.org/12251010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182281 4ff67af0-8c30-449e-8e8b-ad334ec8d88c