Make BrowserList::(Add|Remove)Observer add the observer to every desktop's browser list.
Also make it impossible to call AddObserver on an individual list from outside BrowserList.
This is to make sure the paradigm becomes that the consumer code is in charge of checking for HostDesktopType (should he care) upon receiving an event; allowing AddObserver to be called on an individual list would put burden on the reader to understand whether the events are coming from multiple browser lists or a single one (which seems unecessary for now).

BUG=129187


Review URL: https://chromiumcodereview.appspot.com/12258019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182384 4ff67af0-8c30-449e-8e8b-ad334ec8d88c