History: Fix selecting multiple visits.
Initially the visist IDs for the checkboxes were set as the visits were
retrieved but currently the visits can be displayed in a different order.

Fix this by setting the ID when the visit is displayed, fixing multiple
selection. Add a test to check that in the future.

BUG=170690

Review URL: https://chromiumcodereview.appspot.com/12217015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182470 4ff67af0-8c30-449e-8e8b-ad334ec8d88c