Apply the pref that prohibits deleting browsing and download history.
The settings() method will now return, for each data type, an object with "selected" and "permitted" properties, describing whether it's selected for removal and whether removal is allowed. The remove() method accepts that input format as well as the simpler one it used to take (i.e. a boolean value for each data type); the "permitted" attribute is ignored by remove().

If an API call would remove a prohibited data type, it instead does nothing and returns an error message.

This patch also adds an error message returned if the remove() method receives an invalid value for one or more specified data types. The data types with valid specifications are still removed.

BUG=175419
TEST=browser_tests --gtest_filter=ExtensionBrowsingDataTest.*

Review URL: https://chromiumcodereview.appspot.com/12212111

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182718 4ff67af0-8c30-449e-8e8b-ad334ec8d88c