Fix dependency on scoped_ptr::reset sequencing in SafeBrowsingDatabaseManager
scoped_ptr<T>::reset() currently guarantees that it deletes the old
stored pointer before assigning its argument to the stored pointer.
This is unsafe, because getting the deleter may result in the
destruction of the scoped_ptr<T> itself. unique_ptr<T> addresses this by
assigning its argument to the stored pointer before deleting the old
value of the stored pointer.

Unfortunately, this breaks code that assumes that the value of the
scoped_ptr will not change during scoped_ptr::reset() before destruction
of the old value is complete.

BUG=176091

Review URL: https://codereview.chromium.org/12261028

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@182943 4ff67af0-8c30-449e-8e8b-ad334ec8d88c