Disable "using base::FilePath" on Linux since it now compiles.
This keeps the using for other platforms temporarily. This patch is to "stop
the bleeding" since new conflicts are added every day. The other platforms will
be fixed in followup patches.

This also adds an exception for the MTPD code which has a circular dependency
on FilePath.

Review URL: https://codereview.chromium.org/12282019

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@183043 4ff67af0-8c30-449e-8e8b-ad334ec8d88c