Fix crash in GetDriveFilePropertiesFunction.
Other uses of render_view_host() are directly under ::RunImpl
so they should be safe. Only this occurrence resides in
a posted task.

Perhaps we should factor out the fs_context retrieval to a function
instead of paying attention to the nullness everywhere, but
for the current patch I'd like to keep it simplest for safely merging
to the beta branch.

BUG=176602
TEST=Open a Drive folder with tons of files and close it immediately (before
the list of files is displayed), and check it won't crash.


Review URL: https://chromiumcodereview.appspot.com/12302004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@183102 4ff67af0-8c30-449e-8e8b-ad334ec8d88c