Disable Add/Delete user controls when user is managed.
Disable the Add/Delete user control in the settings page
when the user is managed. The user is still able to modify
his own profile. Remove the check for ManagedMode in
profile_manager.cc which previously was used to do not show
the users section of the settings at all when in managed mode.
Screenshots are here:
Without user profile selected:
https://docs.google.com/a/google.com/file/d/0BzMCIwDgeNezUThxUHdwN2NrRlE/edit?usp=sharing
With user profile selected:
https://docs.google.com/a/google.com/file/d/0BzMCIwDgeNezZnJ1cGFUZEFzR0k/edit?usp=sharing

BUG=171370
TEST=Manual

Review URL: https://chromiumcodereview.appspot.com/12296009

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@183478 4ff67af0-8c30-449e-8e8b-ad334ec8d88c