Fix regression that ChromeOS file chooser ignores "support_gdata" option.
When the option is set to false, "Google Drive" tab must be hidden
from the file selection dialog because the caller does not support
Drive path yet.

The option, however, is ignored in the latest version
since it has been renamed in JavaScript side in rafactoring
work http://crbug.com/162803 but not in the C++ host.
This patch fixes the mismatch.

BUG=176793

Review URL: https://chromiumcodereview.appspot.com/12302041

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@183491 4ff67af0-8c30-449e-8e8b-ad334ec8d88c