Improved redirect handling in the preview/allow flow.
By adding the |new_navigation_| member and listening to the DidUserGetGesture()
we are able to better detect whether the navigation was caused by an user or by
a redirect. This information gets added to and updated on the temporary
exception map.

Also corrected simulating the click on the infobar in the tests by calling the
dismiss function, testing the fact that the Accept and Cancel functions return
true and removing the infobar instead of just accepting.

R=pam@chromium.org, bauerb@chromium.org
BUG=168772
TEST=out/Debug/browser_tests --gtest_filter=ManagedModeBlockModeTest.*


Review URL: https://chromiumcodereview.appspot.com/12327007

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@183562 4ff67af0-8c30-449e-8e8b-ad334ec8d88c