Merge BrowserListImpl into BrowserList.
Kept AddBrowser and RemoveBrowser static (having it access the private members of the appropriate list for the browser added/removed); kept AddObserver and RemoveObserver static so that observers are notified about browser additions, removals, and set actives on every desktop (observers should filter if they so desire).

The only calls that are not-static (i.e. for which consumers need to be aware of HostDesktopType) are: iterators, size(), empty(), get(), and GetLastActive(). All the consumers were already made multi-desktop aware in the multiple CLs preceding this one :)!

BUG=129187


Review URL: https://chromiumcodereview.appspot.com/12281006

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@183580 4ff67af0-8c30-449e-8e8b-ad334ec8d88c