Remove special case warning for Sirefef.
(The public key block is still in place - this just removes the message.)

If the message hasn't worked by now, then it's not going to.  If I ever get
around to putting message support in the CRLSet, then this code will still be
in the repo history.

BUG=none

Review URL: https://codereview.chromium.org/12276004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@183582 4ff67af0-8c30-449e-8e8b-ad334ec8d88c