History: Use images instead of unicode characters.
Unicode characters are font dependent so they may look differently on some
machines. This patch uses images for the arrows in grouping by domain and the
range next and previous buttons.

This patch also sets the domain name for grouped domains to a link so that users can tab to it and and pressing enter will expand.

Uses 2 images that will be uploaded in a different CL.

BUG=170690
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12218058

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@183795 4ff67af0-8c30-449e-8e8b-ad334ec8d88c