Factor out uses of the WebContentsModalDialog interface from platform-independent code
This CL is part 2 of 6 in the work to remove the WebContentsModalDialog
interface. It removes platform-independent uses of functions on the
WebContentsModalDialog interface in the UI code and replaces them with
the platform-specific equivalents.

Patch set 1 removes CloseWebContentsModalDialog call from LoginHandler.

This CL builds on part 1 in https://codereview.chromium.org/12281012.

BUG=157161

Review URL: https://chromiumcodereview.appspot.com/12276010

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@183995 4ff67af0-8c30-449e-8e8b-ad334ec8d88c