fix back/forward keys on win aura
WM_APPCOMMAND of "browser forward/backward" is synthesized by
DefWindowProc when the WM_XBUTTONUP message is passed to it.
Don't unconditionally mark event as handled in
RenderWidgetHostViewAura::OnMouseEvent so that they're passed to
DefWindowProc, which then sends the appropriate WM_APPCOMMAND.
And, change #ifs for APPCOMMAND so that it translates to the IDC_
for back/forward. (Do you remember why that was #if'd out by any
chance?)

R=ben@chromium.org
BUG=175678,177272

Review URL: https://chromiumcodereview.appspot.com/12315048

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184046 4ff67af0-8c30-449e-8e8b-ad334ec8d88c