Do not initialize sync or policy on import process.
This fixes a conflict that occurs when the WebDataService is initialized on both the Import and normal processes
and they both try to open the same DB file.

This is a temporary fix for M26 - for M27 we will change ProfileDependencyManager to not launch any services on the
import process.

TBR=pkasting

BUG=174864


Review URL: https://chromiumcodereview.appspot.com/12310042

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184116 4ff67af0-8c30-449e-8e8b-ad334ec8d88c