Add support for a manager passphrase of a managed account.
The managed user settings page will be shown with every control disabled. A click on the button "Unlock" shows the passphrase dialog. After a successful authentication, the controls of the settings page are enabled. The authentication is reset when the settings page
is closed.

BUG=171370
TEST=browser_tests

Review URL: https://chromiumcodereview.appspot.com/11783008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184206 4ff67af0-8c30-449e-8e8b-ad334ec8d88c