Deleting notifications that don't get updates. If an existing notification is not present in the next batch of cards received from the server, we delete it.
The set activeNotifications is now rebuild from scratch on every batch, which eliminates the risk of infinite growth. This also allowed to NOT delete notifications from this set upon closing the notification, which eliminated a risk of race conditions (for 'activeNotifications' data) between creating and deleting/closing notifications. This also allowed to not wait for successful notification creation before we add notification to the map upon creation, with a similar race conditions-related benefit.

BUG=164227

Review URL: https://chromiumcodereview.appspot.com/12260047

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184225 4ff67af0-8c30-449e-8e8b-ad334ec8d88c