Non-web-accessible extension URLs should not load in non-extension processes
Blocking of URL loads in the browser process cannot be as extensive as the one done in the renderer process, as we don't have the frame URL and page URL at resource load time. I've tried to do a similar check with the data available on the browser side.

The main part which I'm not entirely happy about is the check for DevTools pages. Currently, if an extension has DevTools page, all of its resources can be loaded by a compromised renderer. I think this can be tightened up with a follow up CL, if we find a good way of distinguishing DevTools processes or permissioning those requests.

BUG=173688

Review URL: https://chromiumcodereview.appspot.com/12218064

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184245 4ff67af0-8c30-449e-8e8b-ad334ec8d88c