Register extensions that use storage.managed at the PolicyService.
The PolicyService can optionally fetch policy for extensions, but it needs to
know which extension IDs to fetch policy for. This CL introduces a tracker that
observes the initial extension list and extension loads and unloads, and keeps
the PolicyService in sync.

Additionally, the policies served through the storage.managed namespace will
come from the leveldb cache until the PolicyService has finalized its
initialization, so that extensions don't see two spurious updates on startup:
first with empty policy, and then with the loaded policy.

BUG=163318


Review URL: https://chromiumcodereview.appspot.com/12314036

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184431 4ff67af0-8c30-449e-8e8b-ad334ec8d88c