Add GetAboutResource to DriveServiceInterface.
Drive API v2 has two APIs to retrieve meta data:
1) "GetAboutResourceOperation" for quota, changestamp and root id.
2) "GetApplistOperation" for a list of installed apps.
On the other hand, GDataWAPI has one API to retrieve them (except root id) at once, which is "GetAccountMeatadataOperation".
It is possible to split the call of GetAccountMetadata into two (for about resource and for app list each). So, to fill the gap, GetAccountMetadata will be migrated into GetAboutResource and GetAppList.
This CL just adds the method to the interface and some stub implementation only. The real implementation for each inheriting class and the client migration will be done in following CLs.

BUG=174237
TEST=Ran unit_tests.


Review URL: https://chromiumcodereview.appspot.com/12342008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184711 4ff67af0-8c30-449e-8e8b-ad334ec8d88c