Fix layering violation in content\browser\renderer_host. That directory shouldn't be including web_contents.
It looks like this regressed because we forgot to update the DEPS exclusion after the tab_contents->web_contents rename. I've fixed the non test case, added exclusion for tests, and put a temporary workaround for the media subdirectory.

It seems that the media subdirectory should move to content\browser since it has dependencies on web_contents.
Review URL: https://codereview.chromium.org/12319119

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184735 4ff67af0-8c30-449e-8e8b-ad334ec8d88c