Check length of wallpaper category before using it.
Normally, we should be able to assume that each wallpaper must belong to at
least one category. But this may not be the case if there is a server error/bug.
To add more safety, we should validate the wallpaper category before using
it.


BUG=177766


Review URL: https://chromiumcodereview.appspot.com/12326077

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184763 4ff67af0-8c30-449e-8e8b-ad334ec8d88c