Omnibox: Fix OnEscapeKeyPressed Crash
The associated bug implies a crash on the line
content::WebContents* contents = controller_->GetWebContents();
if (!contents->IsLoading()) {

This crash looks similar to the crash observed in
crbug.com/12250014
where we also assumed that GetWebContents() was non-NULL.

That was fixed by
https://chromiumcodereview.appspot.com/12250014

Peter wrote:
>>>
If we're still getting reports for this crash, then I'd say let's check in and
see if it makes the reports stop.

I'd also like to know if anywhere else assumes a non-NULL WebContents.  If this
makes the crash reports stop, we should fix any other such places too.
>>>

This is one of the other places.

BUG=167510


Review URL: https://chromiumcodereview.appspot.com/12340087

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184845 4ff67af0-8c30-449e-8e8b-ad334ec8d88c