Move the responsibility to update the StateMachine from BurnController to BurnManager.
This is a part of migration from BurnController to BurnManager.
By this CL, BurnManager takes all the responsibility to update the StateMachine.
Note: BurnController still accesses to the StateMachine to check the current
status. The code will be migrated somehow into BurnManager or more structured
code in following CLs.
Note that, this CL changes the order of BurnController::OnBurnStateChanged(INITIAL) and BurnController::OnBurnProgressUpdated(BURN_SUCCESS, ...)
for the success callback. In the old code, the callback trace is:
- BurnManager::BurnProgressUpdated(BURN_SUCCESS)
- BurnController::OnBurnProgressUpdated(BURN_SUCCESS)
- BurnController::FinalizeBurn()
- StateMachine::OnSuccess()
- BurnController::OnBurnStateChanged(INITIAL)
while, in the new code, it is:
- BurnManager::BurnProgressUpdated(BURN_SUCCESS)
- StateMachine::OnSuccess()
- BurnController::OnBurnStateChanged(INITIAL)
- (back to BurnManager::BurnProgressUpdated)
- BurnController::OnBurnProgressUpdated(BURN_SUCCESS)
- BurnController::FinalizeBurn()
However, in real use case, BurnController::OnBurnStateChanged(INITIAL) does nothing and BurnController::OnBurnProgressUpdated just invoke FinalizeBurn. So it shouldn't change the behavior observable from outside of the class.

BUG=126979
TEST=Ran unit_tests and manually burn the image on Chrome OS.


Review URL: https://chromiumcodereview.appspot.com/12317095

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184958 4ff67af0-8c30-449e-8e8b-ad334ec8d88c