Revert 184871
This is really a shot in the dark, so apologies if this turns out to be
an inappropriate revert.

Some file system api tests started failing about ten hours ago on Linux Tests:
FileSystemApiOpenBackgroundTest
FileSystemApiSaveBackgroundTest
http://build.chromium.org/p/chromium.linux/buildstatus?builder=Linux%20Tests%20%28dbg%29%281%29&number=23657

They've been red on a number of platforms ever since.

There was an obvious contender on the blamelist
http://src.chromium.org/viewvc/chrome?view=rev&revision=184878

An earlier sheriff reverted this but the problem did not go away.

So I'm stuck trying to figure out what else could've cause this.  There
are no likely candidates, but this is the most likely in my assessment
and that of my fellow sheriff of an unlikely lot.

So I'm reverting and praying.

> Adds cache invalidation for out of date entries in File Loader extension.
> 
> Image loader was missing cache invalidation for images which became out of date. It means, that once cached they stayed there forever until the cache eviction or a manual removal. This patch solves this problem by adding an additional argument - timestamp, which can be passed to the ImageLoader.Client instance. If the cached image has a different timestamp, then the cache entry will be invalidated and replaced with the new image.
> 
> TEST=Not used in production, yet.
> BUG=178361
> 
> Review URL: https://codereview.chromium.org/12326150

TBR=mtomasz@chromium.org
Review URL: https://codereview.chromium.org/12310173

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184969 4ff67af0-8c30-449e-8e8b-ad334ec8d88c