Improve infobar checking in managed mode browser tests.
Initially we checked the number of infobars shown, but that proved to be flaky
in some cases, so now check that the actual infobars we are expecting are
present or not. The check whether it is the right infobar or not is done by
looking at the infobar message.

R=bauerb@chromium.org, pam@chromium.org
BUG=168772
TEST=out/Debug/browser_tests --gtest_filter=ManagedModeBlockModeTest.*


Review URL: https://chromiumcodereview.appspot.com/12271018

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@184984 4ff67af0-8c30-449e-8e8b-ad334ec8d88c