Added an updater for component policy.
This class downloads policy for extensions from the cloud policy server, and
stores it locally. This is the bottom-half of that service; it downloads the
policy data indicated by the server. The other half contacts the server to get
the information about which extensions have policies, and where they are located.

TBR=jochen@chromium.org
BUG=163318


Review URL: https://chromiumcodereview.appspot.com/12334014

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185033 4ff67af0-8c30-449e-8e8b-ad334ec8d88c