Remove socket_map_.size() == 0 CHECK.
This is a re-landing of https://chromiumcodereview.appspot.com/12320016/ , but simply removing the CHECK instead of using ApiResourceManager. I couldn't use ApiResourceManager because using it in UI thread seems to cause memory leak which is hard to fix. I think it's somehow related to |io_data_| in ExtensionSystem, but I think I should just go with a simpler solution of removing this line.

BUG=135470

Review URL: https://chromiumcodereview.appspot.com/12314138

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185057 4ff67af0-8c30-449e-8e8b-ad334ec8d88c