Revert 185029
We see a ton of browser_tests timeouts on windows.
http://build.chromium.org/p/chromium.win/builders/Win7%20Tests%20%281%29/builds/20141

I notice this change didn't run browser_tests on any non-Aura windows boxes.
And it removes a keyboard item from windows.

Working theory:
maybe some browsers tests are sending that particular key that isn't around any more on windows and waiting for something to happen and waiting forever.

Reverting to test the theory.

Apologies if the theory turns out wrong. :-|


> Remove the Exit menu item from the Wrench menu in Win/Ash. 
> 
> Relanding after a revert in crrev.com/184662 due to accidentally disabling the Exit menu on non Win-Ash platforms. 
> 
> This patch is identical to patch set 4 in https://codereview.chromium.org/12330056/
> 
> BUG=176130
> TEST=Metro/Ash chrome browsers lose their Exit menu item.
> TBR=cpu, sky
> 
> Review URL: https://chromiumcodereview.appspot.com/12340110

TBR=robertshield@chromium.org

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185061 4ff67af0-8c30-449e-8e8b-ad334ec8d88c