Require manifests for native messaging hosts.
Previously native messaging hosts were discovered by looking in the user 
data directory. That approach is hard to use and is not reliable. Now 
chrome will look for a host manifest file in a system-specific directory 
and load that file to find host binary.

BUG=142915
TBR=thakis@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12285015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185189 4ff67af0-8c30-449e-8e8b-ad334ec8d88c