Remove uses of ConstrainedWindowViews within Views dialogs
This change moves towards the elimination of the ConstrainedWindowViews
type by replacing ConstrainedWindowViews-specific functions with their
Widget-based equivalents, removing the ConstrainedWindowViews-specific
functions, and updating web contents modal dialog users to reference
the window via the views::Widget type.

Still remaining to be done for Views is the removal of the
ConstrainedWindowViews::Create factory function.

Patch set 1 is the removal of the ConstrainedWindowViews-specific functions

BUG=157161


Review URL: https://chromiumcodereview.appspot.com/12389003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185216 4ff67af0-8c30-449e-8e8b-ad334ec8d88c