Split knowledge of WebContents out of OmniboxEditModel.
Separates the elements of OmniboxEditModel that interact with the current page via WebContents into a new OmniboxCurrentPageDelegate.

The goal of this change is to support OmniboxEditModel usage on iOS, which
doesn't use WebContents. An iOS implementation of OmniboxCurrentPageDelegate will follow at a later point.


Review URL: https://chromiumcodereview.appspot.com/12304023

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185289 4ff67af0-8c30-449e-8e8b-ad334ec8d88c