Require manifests for native messaging hosts.
Previously native messaging hosts were discovered by looking in the user
data directory. That approach is hard to use and is not reliable. Now
chrome will look for a host manifest file in a system-specific directory
and load that file to find host binary.

This is a second attempt to land this CL as it was reverted (see
https://codereview.chromium.org/12386018). The problem that caused it to
be reverted was not related to this CL.

BUG=142915
TBR=mpcomplete@chromium.org

Review URL: https://codereview.chromium.org/12389041

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185382 4ff67af0-8c30-449e-8e8b-ad334ec8d88c