Extract DriveURLRequestJob to a brand new file.
This is a preparation to write unit tests for the DriveURLRequestJob.
Currently, it is contained by the anonymous namespace. This CL move it to a
new file with resolving some namespaces and include directive.
This CL doesn't contain test cases. These will be added in following CLs.

BUG=159383
TEST=Ran unit_tests and invoke drive: protocol manually.
TBR=thestig@chromium.org for chrome/chrome_browser_chromeos.gypi

Review URL: https://chromiumcodereview.appspot.com/12387016

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185488 4ff67af0-8c30-449e-8e8b-ad334ec8d88c