History: Add lock/unlock button to the history page.
These buttons will only be visible if the profile of the
user is managed. The Unlock button is visible if the managed
user is currently not in elevated state, the Lock button is
visible if the managed user is in elevated state.

Clicking such a button toggles the elevation state. Clicking
the unlock button opens a passphrase dialog, and only if the
authentication is successful, the managed user is set in
elevated state.

Here are screenshots for a managed user in elevated state
(history_unlocked.png) and not in elevated state
(history_locked.png):
https://drive.google.com/a/google.com/folderview?id=0BzMCIwDgeNezazRYa3lpbkxCdnM&usp=sharing

BUG=171370
TEST=Manual

Review URL: https://chromiumcodereview.appspot.com/12340004

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185505 4ff67af0-8c30-449e-8e8b-ad334ec8d88c