files: Stop computing a directory size when selected
Selecting a directory was expensive because Files.app computed the size of
a directory by recursively visiting all the sub directories and summing up
sizes of all files underneath. This is super expensive with a large directory
tree with thousands of directories and files underneath.

BUG=179000
TEST=Select a directory, size is not shown. Select a file, size is shown.


Review URL: https://chromiumcodereview.appspot.com/12379037

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@185527 4ff67af0-8c30-449e-8e8b-ad334ec8d88c