Fix pre-allocation in WebRequestConditionAttributeResourceType::Create
In WebRequestConditionAttributeResourceType::Create, the vector of types was created via the "resize" constructor to the desired final size, but then items were added again using push_back. The added items were never read, and the default 0 values were used on their place.

This CL changes the implicit "resize" to an explicit "reserve", and also contains a minor fix in the unit tests, as well as a modification of the tests to catch this kind of error in the future.


BUG=179505


Review URL: https://chromiumcodereview.appspot.com/12429005

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@186183 4ff67af0-8c30-449e-8e8b-ad334ec8d88c