Implement Windows accessibility for Aura.
Windows accessibility requires posting notifications
on an HWND, then listening for messages on that HWND
and routing them back to the appropriate child object.
Before, that code was implemented in NativeWidgetWin
because a NativeWidget corresponded to an HWND.

Now with Aura, a Widget doesn't necessarily have an
HWND, so there's now a new class,
NativeWidgetAccessibilityWin, that's a companion to
NativeViewAccessibilityWin and handles this abstraction.

Refactored some code so that we don't need screen
reader detection or dynamic HWND annotation (in
methods like SetAccessibilityRole, etc.) anymore.

Tested with Windows screen readers. Web content is
still not accessible in Aura, that will be the next
changelist. Win 8 in Metro mode is not accessible yet,
that's going to require further changes.

BUG=175156


Review URL: https://chromiumcodereview.appspot.com/12335075

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@186382 4ff67af0-8c30-449e-8e8b-ad334ec8d88c