Implement CopyHostedDocument on Drive API v2.
This is the implementation of the CopyHostedDocument.
The new operation class CopyResourceOperation is not limited to the hosted
document. It should work with the regular files, too, conceptually.
However, it takes only needed parameters to support CopyHostedDocument, and
should be going to be extended in following CLs.

BUG=179838
TEST=Ran unit_tests


Review URL: https://chromiumcodereview.appspot.com/12432003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@186396 4ff67af0-8c30-449e-8e8b-ad334ec8d88c