History: Create delete directives when local or remote history is deleted.
In order to support offline deletion of synced history entries, create a
delete directive whenever entries are deleted from the history page. The
directives will be synced to other clients, ensuring that it's possible
to delete remove history entries even when the history server cannot be
reached. This CL only creates delete directives when specific visits are
deleted, not for time range deletions from the Clear Browsing Data dialog.

This also adds a new HistoryEntry type, which can represent either a local
or remote history entry. Compared to using DictionaryValue, this is a much
better format for processing the entries before they are returned to the
history front-end. This will also make it easier to push the parsing of
the history server's JSON response into web_history_service.cc.

BUG=171693

Review URL: https://codereview.chromium.org/12319093

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@186400 4ff67af0-8c30-449e-8e8b-ad334ec8d88c