FakeDriveService should distinguish root entries and orphan entries.
- RemoveFromDirectory from the root folder is an error if a parent link did not exist.
- AddNewDirectory to the root folder creates a parent link.
- InitialUpload to the root folder creates a parent link.
- Unit tests and the test data changed accordingly.

Some notes:
* The real GDataWapiService still mixes root entries and orphan
entries, since we have not flagged "showroot=true" yet. But we are
going to flag it on soon. The current patch can be considered as a
pre-flight test that existing code already handling the disambiguation well.
* Test feeds other thatn root_feed.json is not touched because they
are not used from FakeDriveService. We should also modify them eventually...

BUG=171207

Review URL: https://chromiumcodereview.appspot.com/12313096

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@186622 4ff67af0-8c30-449e-8e8b-ad334ec8d88c