Add back a DCHECK on threading behavior in HostContentSettingsMap.
HostContentSettingsMap::RegisterExtensionService is the final piece of
initialization in this class, after which content_settings_providers_
is immutable, and so can be used from multiple threads without
locking. This assert verifies that the member is only accessed from a
single thread before initialization completes. The previous assert
(documented in the bug) checked that it was never read before
initialization, which was incorrect, it does get read but only by the
same thread that constructs and initializes it.

BUG=176315

Review URL: https://codereview.chromium.org/12553003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@186777 4ff67af0-8c30-449e-8e8b-ad334ec8d88c