Made notification center notifications collapsible and expandable.
Implemented collapse/expand for notifications in the notification
center. Notifications are collapsed by default and can be expanded by
pressing their expand button. Once expanded, they can't be collapsed
again.

Limitations of this version are:
1) Notifications in alerts/toasts are not affected (they're always expanded).
2) Notifications get expand buttons even if they have no content to expand.
3) Collapsed text is always collapsed to 1 lines (not 2 or 3 as required) and expanded text
is always expanded to its full height (not limited to 7 lines as
required).
4) Notification expansion is lost when notifications are
added to the notification center or when the notification center is
closed and opened again. These issues will be addressed in separate
change lists.

BUG=161098
TBR=stevenjb@chromium.org

Review URL: https://chromiumcodereview.appspot.com/12326091

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@186873 4ff67af0-8c30-449e-8e8b-ad334ec8d88c