Initial UI structure for autofill dialog
This is the basic UI structure that has some of the trnasitions and
animations build in. The dialog is only initialized only in related
autofill code and this is mainly to get a UI structure committed to
build around to. The placeholders will be removed and some animations
and control structures may eventually change.

BUG=177380
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/12412003

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187071 4ff67af0-8c30-449e-8e8b-ad334ec8d88c