Update TestURLRequest constructor interface in extensions unit tests.
The TestURLRequest constructor interface is being expanded to allow separate
initialization arguments for the URLRequestContext and NetworkDelegate.
Most of the time this means making it explicit that the test in question
uses no network delegate.  This is motivated by a desire to remove external
accesses to a request context's delegate.  (If a block of code needs to
use a NetworkDelegate, it should not query the URLRequestContext for it,
but instead should already have access to whatever delegate the context uses.)

BUG=146587


Review URL: https://chromiumcodereview.appspot.com/12328079

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187192 4ff67af0-8c30-449e-8e8b-ad334ec8d88c