Remove the action choose dialog's icon and title.
By creating a browser and later adding a tab using chrome::AddSelectedTabWithURL it is impossible to get rid of the icon and the title.
However, chrome::OpenApplication can do that, but it doesn't allow to pass dimensions. Therefore, an argument has been added to be able to override bounds.

TEST=Insert an sd card with a DCIM folder. Verify the action choose dialog.
BUG=179084

Review URL: https://chromiumcodereview.appspot.com/12491008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187236 4ff67af0-8c30-449e-8e8b-ad334ec8d88c