Fixed a glitch in the ComponentCloudPolicyUpdater where a protobuf field was read before being validated.
The updater receives protobufs that contain a secure hash of some content. When this hash
differs from the cached version then the content has been updated. The bug is that the
updater compares an empty secure_hash with an empty cached hash, which causes an early
return but should cancel a pending job, if there is one.

BUG=163318


Review URL: https://chromiumcodereview.appspot.com/12542015

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187290 4ff67af0-8c30-449e-8e8b-ad334ec8d88c