Regularize ownerships and lifecycle for storage monitor platform classes.
This change regularizes the construction lifecycle of the StorageMonitor implementations so that they always exist for the construction of the Profile, but don't issue any notifications until after that construction is complete. This allows profile-based listeners to be sure they're synchronized with the monitor.

It also regularizes ownership of all storage-monitor-related objects. The base platform-specific object is owned by the ChromeBrowserMain object, and owns all subsidiary platform-specific objects needed to monitor and/or deal with volumes of various types (i.e. MTP).

R=vandebo@chromium.org
BUG=None


Review URL: https://chromiumcodereview.appspot.com/12334096

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187332 4ff67af0-8c30-449e-8e8b-ad334ec8d88c