Remove calls to URLRequestContext::network_delegate() in chrome browser code.
These changes are motivated by a desire to remove external accesses to a request
context's delegate. (If a block of code needs to use a NetworkDelegate, it should
not query the URLRequestContext for it, but instead should already have access
to whatever delegate the context uses.)  Two interfaces changed to support this:

ProxyServiceFactory::CreateProxyService() now accepts separate arguments for
URLRequestContext* and NetworkDelegate*, and its call locations (most notably
from chrome/browser/io_thread.cc) specify the network delegate the context used.
Note that this means that if someone gets a non-const URLRequestContext pointer
and later calls set_network_delegate() on it, the ProxyServiceFactory would not
match the context's delegate.  This doesn't happen now, and I can't even think of
why we might want to do that, but it's worth noting for the record that this would
be a Bad Thing(tm).

The constructor of TestURLRequest also accepts a separate argument for the
network delegate (as an alternative to using the context's delegate).  Several
tests were updated to clarify their url requests have no network delegate.

BUG=146587


Review URL: https://chromiumcodereview.appspot.com/12313094

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187336 4ff67af0-8c30-449e-8e8b-ad334ec8d88c