Mac: Fix bookmark bar context menu crash.
The problem was caused because the BMB controller would create the context 
menu controller in -loaded:, which could be called multiple times - which made 
the weak refs to the menu controller invalid. 

This change instead creates the context menu controller only once in -init... 
and validates that the bookmark model is valid when asked for a menu. 

BUG=181547
TEST=Open up the WebUI bookmarks manager and rename a folder. Now,
right-click on the BMB to bring up the context menu from the bookmark bar.
It should not crash.

Review URL: https://chromiumcodereview.appspot.com/12755008

git-svn-id: http://src.chromium.org/svn/trunk/src/chrome/browser@187448 4ff67af0-8c30-449e-8e8b-ad334ec8d88c